break-into-modules: rename commons.essentials to commons.core
