+++ apigateway/src/main/java/nva/commons/apigateway/ApiMessageParser.java
@@ -8,8 +8,8 @@ import java.io.IOException;
 import java.io.StringReader;
 import java.util.Optional;
 import nva.commons.apigateway.exceptions.ApiIoException;
-import nva.commons.essentials.JacocoGenerated;
-import nva.commons.essentials.JsonUtils;
+import nva.commons.commons.JacocoGenerated;
+import nva.commons.commons.JsonUtils;
 
 /**
  * Not intended for use outside the {@link RestRequestHandler}. Class for parsing a message from ApiGateway.
