+++ apigateway/src/main/java/nva/commons/apigateway/AuthorizedApiGatewayHandler.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 import nva.commons.apigateway.exceptions.ApiGatewayException;
-import nva.commons.essentials.Environment;
+import nva.commons.commons.Environment;
 import org.slf4j.Logger;
 
 public abstract class AuthorizedApiGatewayHandler<I, O> extends ApiGatewayHandler<I, O> {
