+++ apigateway/src/test/java/nva/commons/apigateway/TestAuthorizedHandler.java
@@ -2,7 +2,8 @@ package nva.commons.apigateway;
 
 import com.amazonaws.services.securitytoken.AWSSecurityTokenService;
 import java.net.HttpURLConnection;
-import nva.commons.essentials.Environment;
+import nva.commons.commons.Environment;
+import org.apache.http.HttpStatus;
 import org.slf4j.LoggerFactory;
 
 public abstract class TestAuthorizedHandler extends AuthorizedApiGatewayHandler<Void, String> {
