+++ core/src/main/java/nva/commons/commons/JacocoGenerated.java
@@ -1,4 +1,4 @@
-package nva.commons.essentials;
+package nva.commons.commons;
 
 public @interface JacocoGenerated {
 }
