+++ core/src/test/java/nva/commons/commons/JsonSerializableTest.java
@@ -1,6 +1,6 @@
-package nva.commons.essentials;
+package nva.commons.commons;
 
-import static nva.commons.essentials.JsonUtils.objectMapper;
+import static nva.commons.commons.JsonUtils.objectMapper;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
