+++ core/src/test/resources/jsonutils/emptyStringValues.json
