+++ core/src/test/resources/jsonutils/missingStringValues.json
