+++ secrets/build.gradle
@@ -1,6 +1,6 @@
 dependencies{
-    compileOnly project(":essentials")
-    testImplementation project(":essentials")
+    compileOnly project(":core")
+    testImplementation project(":core")
 
     compileOnly project(":apigateway")
     testImplementation project(":apigateway")
